fix: fsd architecture
