+++ src/features/levelup-feedback/index.ts
@@ -8,3 +8,7 @@ export { CardInfo } from './ui/CardInfo'
 export { useFeedbackData } from './model/useFeedbackData'
 export { getAttemptDetails } from './api/getAttemptDetails'
 export { getCardDetails } from './api/getCardDetails'
+export { deleteAttempt } from './api/deleteAttempt'
+export { FeedbackLoader } from './ui/FeedbackLoader'
+export type { FeedbackItem, FeedbackProcessingStep, FeedbackStatus } from './model/feedbackTypes'
+export { useFeedbackPolling } from './model/useFeedbackPolling'
